+++ .github/ISSUE_TEMPLATE/⚙️-프로젝트.md
@@ -1,0 +1,14 @@
+---
+name: "⚙️ 프로젝트"
+about: 빌드 설정, 환경 설정, CI/CD 스크립트 등 프로젝트 관리에 관련된 변경
+title: '프로젝트 : '
+labels: documentation
+assignees: brobro332
+
+---
+
+## 🎈 개요
+
+## 📑 본문
+
+## 🚀 기타
