+++ .github/ISSUE_TEMPLATE/💎-리팩토링.md
@@ -1,0 +1,14 @@
+---
+name: "\U0001F48E 리팩토링"
+about: 동작에는 영향을 주지 않지만, 코드 구조나 품질을 개선하는 경우
+title: '리팩토링 : '
+labels: documentation
+assignees: brobro332
+
+---
+
+## 🎈 개요
+
+## 📑 본문
+
+## 🚀 기타
