+++ .github/ISSUE_TEMPLATE/🔥-기능추가.md
@@ -1,0 +1,14 @@
+---
+name: "\U0001F525 기능추가"
+about: 새로운 기능을 추가하는 경우
+title: '기능추가 : '
+labels: documentation, enhancement
+assignees: brobro332
+
+---
+
+## 🎈 개요
+
+## 📑 본문
+
+## 🚀 기타
