+++ .github/ISSUE_TEMPLATE/🔧-기능변경.md
@@ -1,0 +1,14 @@
+---
+name: "\U0001F527 기능변경"
+about: 기능은 유지되지만 로직이나 조건이 바뀌는 경우
+title: '기능변경 : '
+labels: documentation
+assignees: brobro332
+
+---
+
+## 🎈 개요
+
+## 📑 본문
+
+## 🚀 기타
