+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,25 @@
+## 📝 설명
+<!-- 어떤 작업을 했는지 간략히 설명해주세요 -->
+
+-
+
+---
+
+## ✅ 체크리스트
+<!-- 코드가 정상적으로 작동하는지 확인함 -->
+
+- [ ]
+
+---
+
+## 📂 관련 이슈
+<!-- 이 PR이 관련된 이슈 번호를 연결해주세요 (예: Closes #42) -->
+
+Closes #
+
+---
+
+## 📌 기타 참고사항
+<!-- 코드 리뷰어가 참고하면 좋을 만한 내용을 추가해주세요 -->
+
+-
