+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'samsami-hub'
+rootProject.name = 'blueprint'
