+++ src/main/java/kr/co/blueprint/BlueprintApplication.java
@@ -1,0 +1,13 @@
+package kr.co.blueprint;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BlueprintApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BlueprintApplication.class, args);
+	}
+
+}
