+++ src/main/java/kr/co/blueprint/domain/model/MessageType.java
@@ -1,0 +1,5 @@
+package kr.co.blueprint.domain.model;
+
+public enum MessageType {
+    ENTER, CHAT, LEAVE
+}
