--- src/main/java/kr/co/samsami_hub/SamsamiHubApplication.java
@@ -1,13 +1,0 @@
-package kr.co.samsami_hub;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SamsamiHubApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SamsamiHubApplication.class, args);
-	}
-
-}
