--- src/main/java/kr/co/samsami_hub/common/config/CorsConfig.java
@@ -1,13 +1,0 @@
-package kr.co.samsami_hub.common.config;
-
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-public class CorsConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-            .allowedOrigins("*")
-            .allowedMethods("*");
-    }
-}
