--- src/main/java/kr/co/samsami_hub/websocket/MessageType.java
@@ -1,5 +1,0 @@
-package kr.co.samsami_hub.websocket;
-
-public enum MessageType {
-    ENTER, CHAT, LEAVE
-}
