--- src/main/java/kr/co/samsami_hub/websocket/repository/MessageRepository.java
@@ -1,7 +1,0 @@
-package kr.co.samsami_hub.websocket.repository;
-
-import kr.co.samsami_hub.websocket.Message;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface MessageRepository extends MongoRepository<Message, Long> {
-}
