+++ src/test/java/kr/co/blueprint/BlueprintApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.blueprint;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BlueprintApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
