--- src/test/java/kr/co/samsami_hub/SamsamiHubApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.co.samsami_hub;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SamsamiHubApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
